% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makef.R
\name{makef}
\alias{makef}
\title{Make non-linear features}
\usage{
makef(x, r, df = 4, tol = 0.01, removeLin = T)
}
\arguments{
\item{x}{Input vector of length \code{nobs}.}

\item{r}{Vector of residuals.}

\item{df}{Degrees of freedom for the fit. Default is 4.}

\item{tol}{A tolerance for same-ness or uniqueness of the x values. To be
passed to the \code{smooth.spline()} function. Default is \code{0.01}.}

\item{removeLin}{If \code{TRUE} (default), removes the linear component from
the newly created non-linear features.}
}
\value{
A list:
\item{f}{Non-linear feature associated with \code{x}.}
\item{nl_predictor}{A function which, when given new data \code{newx}, returns 
the value of the non-linear predictor at those \code{x} values. Needed for 
creating the non-linear features for new data.}
}
\description{
Internal function for making non-linear features.
}
